requirements.txt
streamlit
langchain-google-genai
python-dotenv
google-generativeai
langchain